+++ Utilities.h
@@ -7,11 +7,11 @@ using namespace std;
 string MoneyToString(long long money);
 
 /* Converts weight in grams into a string formatted as 12.345 (kilos.grams) */
-string WeightToString(size_t weight); 
+string WeightToString(long long weight);
 
 /* Takes date/time in a form of tm structure and returns a string with the date formatted as DD/MM/YYYY.
  * If showTime is true, adds HH:MM:SS */
 string TmToString(tm date, bool showTime = false);
 
 /* Adds filler character to the end of the string if its length is less that width argument */
-string SetStringWidth(string input, size_t width, bool adjustRight = false, char filler = ' ');
+string SetStringWidth(string input, int width, bool adjustRight = false, char filler = ' ');
